.history/requirements_20250920005934.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
gunicorn>=20.1.0
pdfplumber>=0.7.5
python-docx>=0.8.11
pydantic>=1.10.0
